Refactor link and title resolver to use actors
This code is over-complicated due to trying to use Java collections. As
part of a general refactor into an Akka-approach, change the link
resolver and title resolver to use actors. We are no longer loading
images for simplicity.

This also allows us to switch out the concurrent hash map for a LRU Map,
which should not continue to endlessly grow but throw out old uris and
titles once they've not been accessed for a while. This is part of the
general refactor to make it so that this can run as a server without
memory issues.